--- schemas/contactsAdd.js
@@ -1,9 +1,0 @@
-const Joi = require('joi');
-
-const contactsAddSchema = Joi.object({
-  name: Joi.required(),
-  email: Joi.string().email().required(),
-  phone: Joi.required(),
-});
-
-module.exports = contactsAddSchema;
