--- schemas/index.js
@@ -1,7 +1,0 @@
-const contactsAddSchema = require('./contactsAdd');
-const contactsUpdateSchema = require('./contactsUpdateSchema');
-
-module.exports = {
-  contactsAddSchema,
-  contactsUpdateSchema,
-};
